requirements.txt
beautifulsoup4
pandas
numpy
matplotlib
markdown
pyyaml
